refactor: move packages to net.thechance.identity instead of identity feature: add IdentityControllerAdvice for exception handling
