+++ identity/src/main/kotlin/net/thechance/identity/api/dto/AuthResponse.kt
