+++ identity/src/main/kotlin/net/thechance/identity/security/handler/AuthErrorResponder.kt
@@ -1,7 +1,7 @@
 package net.thechance.identity.security.handler
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import identity.security.handler.ApiErrorResponse
+import net.thechance.identity.security.handler.ApiErrorResponse
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.stereotype.Component
